refactor: 매퍼 클래스를 단일 정적 내부 클래스로 통합
