+++ src/main/java/com/hhpl/concertreserve/interfaces/api/controller/ConcertController.java
@@ -2,12 +2,12 @@ package com.hhpl.concertreserve.interfaces.api.controller;
 
 import com.hhpl.concertreserve.application.facade.ConcertFacade;
 import com.hhpl.concertreserve.application.facade.WaitingQueueFacade;
-import com.hhpl.concertreserve.application.model.ConcertInfo;
-import com.hhpl.concertreserve.application.model.ReservationInfo;
-import com.hhpl.concertreserve.application.model.ScheduleInfo;
-import com.hhpl.concertreserve.application.model.SeatInfo;
+import com.hhpl.concertreserve.application.model.concert.ConcertInfo;
+import com.hhpl.concertreserve.application.model.concert.ReservationInfo;
+import com.hhpl.concertreserve.application.model.concert.ScheduleInfo;
+import com.hhpl.concertreserve.application.model.concert.SeatInfo;
 import com.hhpl.concertreserve.interfaces.api.common.ApiResponse;
-import com.hhpl.concertreserve.interfaces.api.mapper.ConcertMapperController;
+import com.hhpl.concertreserve.interfaces.api.mapper.ControllerMapper;
 import com.hhpl.concertreserve.interfaces.dto.concert.*;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
@@ -21,13 +21,12 @@ import java.util.List;
 public class ConcertController {
     private final WaitingQueueFacade waitingQueueFacade;
     private final ConcertFacade concertFacade;
-    private final ConcertMapperController concertMapper;
 
     @Operation(summary = "콘서트 조회 API", description = "예매 가능한 콘서트 목록을 조회한다.")
     @GetMapping("/")
     public ApiResponse<List<ConcertResponse>> getAvailableConcerts() {
         List<ConcertInfo> concerts = concertFacade.getAvailableConcerts();
-        List<ConcertResponse> response = concertMapper.toConcertResponseList(concerts);
+        List<ConcertResponse> response = ControllerMapper.ConcertMapper.toConcertResponseList(concerts);
         return ApiResponse.OK(response);
     }
 
@@ -36,7 +35,7 @@ public class ConcertController {
     public ApiResponse<List<ScheduleResponse>> getAvailableSchedules(@PathVariable Long concertId,@RequestHeader("X-QUEUE-UUID") String uuid) {
         waitingQueueFacade.validateQueueStatusForUser(uuid,concertId);
         List<ScheduleInfo> schedules = concertFacade.getAvailableSchedules(concertId);
-        List<ScheduleResponse> response = concertMapper.toScheduleResponseList(schedules);
+        List<ScheduleResponse> response = ControllerMapper.ConcertMapper.toScheduleResponseList(schedules);
         return ApiResponse.OK(response);
     }
 
@@ -47,7 +46,7 @@ public class ConcertController {
                                                              @RequestHeader("X-QUEUE-UUID") String uuid) {
         waitingQueueFacade.validateQueueStatusForUser(uuid,concertId);
         List<SeatInfo> seats = concertFacade.getAvailableSeats(scheduleId);
-        List<SeatResponse> response = concertMapper.toSeatResponseList(seats);
+        List<SeatResponse> response = ControllerMapper.ConcertMapper.toSeatResponseList(seats);
         return ApiResponse.OK(response);
     }
 
@@ -58,7 +57,7 @@ public class ConcertController {
                                                          @RequestBody ReservationRequest request) {
         waitingQueueFacade.validateQueueStatusForUser(uuid,concertId);
         ReservationInfo reservation = concertFacade.createTemporarySeatReservation(uuid, request.seatId());
-        ReservationResponse response = concertMapper.toReservationResponse(reservation);
+        ReservationResponse response = ControllerMapper.ConcertMapper.toReservationResponse(reservation);
         return ApiResponse.OK(response);
     }
